Add knight, bishop and rook attacks

+++ src/bitboard/test/bboardtest.cpp
@@ -10,12 +10,10 @@ CPPUNIT_TEST_SUITE_REGISTRATION(BBoardTest);
 BBoardTest::BBoardTest()
   : CppUnit::TestFixture()
 {
-
 }
 
 BBoardTest::~BBoardTest()
 {
-
 }
 
 void BBoardTest::test_create()
@@ -24,25 +22,21 @@ void BBoardTest::test_create()
   U64 black = C64(0xAA55AA55AA55AA55);
   U64 white = ~black;
   Print(white);
-  CPPUNIT_ASSERT_EQUAL(Uint8(1), LSBIndex(white));
-
+  CPPUNIT_ASSERT_EQUAL(1, LSBIndex(white));
   U64 lsb[64];
-  for (Uint8 i = 0; i < 64; i++)
+  for (int i = 0; i < 64; i++)
   {
     lsb[i] = C64(0);
     lsb[i] |= (C64(1) << i);
     CPPUNIT_ASSERT_EQUAL(i, LSBIndex(lsb[i]));
   }
-
   CPPUNIT_ASSERT_EQUAL(32, CountBits(black));
 }
 
 void BBoardTest::setUp()
 {
-
 }
 
 void BBoardTest::tearDown()
 {
-
 }
